SQUASHME: pnfs: fix pnfs_commit update_layout range
Whole file semantics are different for COMMIT (0,0) and layouts
(0,NFS4_MAX_UINT64).

Reported-by: Alexandros Batsakis <batsakis@netapp.com>

Translate COMMIT whole file semantics to LAYOUTGET whole file semantics.

The LAYOUTGET request uses u64 length, change pnfs_update_layout() count field
type to match.

Replace PAGE_SIZE with PAGE_CACHE_SIZE and PAGE_CACHE_SHIFT.

+++ fs/nfs/pnfs.h
@@ -34,7 +34,7 @@ extern int pnfs4_proc_layoutreturn(struct nfs4_pnfs_layoutreturn *lrp);
 extern const nfs4_stateid zero_stateid;
 
 int pnfs_update_layout(struct inode *ino, struct nfs_open_context *ctx,
-	size_t count, loff_t pos, enum pnfs_iomode access_type,
+	u64 count, loff_t pos, enum pnfs_iomode access_type,
 	struct pnfs_layout_segment **lsegpp);
 
 int _pnfs_return_layout(struct inode *, struct nfs4_pnfs_layout_segment *,
